refactor(sdk)!: move HTTP types to the helpers http submodule
HttpAuthConfig, HttpInvocationConfig, HttpMethod, the http helper, and the
buffered HttpRequest/HttpResponse now live only in @iii-dev/helpers/http and are
removed from the root SDK public surface (no re-export, no Api* aliases). The
buffered types reclaim the Http* names; the old streaming Http*=Stream* aliases
and the Api* names are gone. The root SDK imports these internally where needed.
