+++ sdk/packages/node/helpers/package.json
@@ -33,6 +33,7 @@
     "type-check": "tsc --noEmit"
   },
   "devDependencies": {
+    "@types/node": "^24.10.1",
     "@vitest/coverage-v8": "^2.1.0",
     "tsdown": "^0.21.4",
     "typescript": "^5.9.3",
