+++ sdk/packages/node/iii/src/index.ts
@@ -14,8 +14,6 @@ export { EngineFunctions, EngineTriggers } from './iii-constants'
 export type {
   AuthInput,
   AuthResult,
-  HttpAuthConfig,
-  HttpInvocationConfig,
   MessageType,
   MiddlewareFunctionInput,
   OnFunctionRegistrationInput,
@@ -34,8 +32,6 @@ export type {
 export type { Trigger, TriggerConfig, TriggerHandler } from './trigger'
 
 export type {
-  ApiRequest,
-  ApiResponse,
   IIIClient,
   InternalHttpRequest,
   RegisterFunctionInput,
@@ -50,10 +46,5 @@ export type {
 /** @deprecated Renamed to `IIIClient`. */
 export type { ISdk } from './types'
 
-/** @deprecated `HttpRequest`/`HttpResponse` renamed to `StreamRequest`/`StreamResponse`. */
-export type { HttpRequest, HttpResponse } from './types'
-
 /** @deprecated Import runtime types from `iii-sdk/runtime`. */
 export type { FunctionRef, TriggerTypeRef } from './runtime'
-
-export { http } from './utils'
