+++ sdk/packages/python/iii-example/pyproject.toml
@@ -10,6 +10,7 @@ authors = [{ name = "III" }]
 requires-python = ">=3.10"
 dependencies = [
     "iii-sdk",
+    "iii-helpers",
     "iii-observability",
     "opentelemetry-api",
     "opentelemetry-sdk",
@@ -17,6 +18,7 @@ dependencies = [
 
 [tool.uv.sources]
 iii-sdk = { path = "../iii", editable = true }
+iii-helpers = { path = "../helpers", editable = true }
 iii-observability = { path = "../observability", editable = true }
 
 [project.scripts]
