+++ sdk/packages/python/iii/tests/test_stream_types.py
@@ -1,6 +1,8 @@
-"""StreamRequest / StreamResponse rename, with Http* back-compat aliases."""
+"""StreamRequest / StreamResponse streaming types and buffered Http* types."""
 
-from iii import HttpRequest, HttpResponse, StreamRequest, StreamResponse
+from iii_helpers.http import HttpRequest, HttpResponse
+
+from iii import StreamRequest, StreamResponse
 
 
 def test_stream_types_exported() -> None:
@@ -8,6 +10,6 @@ def test_stream_types_exported() -> None:
     assert StreamResponse is not None
 
 
-def test_http_names_alias_stream_types() -> None:
-    assert HttpRequest is StreamRequest
-    assert HttpResponse is StreamResponse
+def test_http_names_are_buffered_types() -> None:
+    assert HttpRequest is not StreamRequest
+    assert HttpResponse is not StreamResponse
