+++ sdk/packages/rust/helpers/Cargo.toml
@@ -17,4 +17,5 @@ path = "src/lib.rs"
 
 [dependencies]
 serde = { version = "1", features = ["derive"] }
+serde_json = { version = "1", features = ["unbounded_depth"] }
 schemars = "0.8"
