+++ sdk/packages/rust/iii-example/Cargo.toml
@@ -5,6 +5,7 @@ edition = "2024"
 
 [dependencies]
 iii-sdk = { path = "../iii" }
+iii-helpers = { workspace = true }
 iii-observability = { workspace = true }
 async-trait = "0.1"
 reqwest = { version = "0.12", default-features = false, features = ["json", "rustls-tls"] }
